+++ bin/postman-to-k6.js
@@ -19,9 +19,10 @@ program
   .option('-i, --iterations <count>', 'Number of iterations.')
   .option('-g, --global <path>', 'JSON export of global variables.')
   .option('-e, --environment <path>', 'JSON export of environment.')
+  .option('--cli-options-file <path>','postman-to-k6 CLI options file. Useful for CI/CD integrations.')
   .option('-c, --csv <path>', 'CSV data file. Used to fill data variables.')
   .option('-j, --json <path>', 'JSON data file. Used to fill data variables.')
-  .option('--k6-params <path>', 'K6 param options config file. Used to set the K6 params used during HTTP requests.')
+  .option('--k6-params <path>','K6 param options config file. Sets K6 params used during HTTP requests.')
   .option('--skip-pre', 'Skips pre-request scripts')
   .option('--skip-post', 'Skips post-request scripts')
   .option('--oauth1-consumer-key <value>', 'OAuth1 consumer key.')
@@ -37,7 +38,7 @@ program
   .option('--oauth1-version <value>', 'OAuth1 version.')
   .option('--oauth1-realm <value>', 'OAuth1 realm.')
   .option('-s, --separate', 'Generate a separate file for each request.')
-  .option('--k6-handle-summary-json <path>', 'Output the K6 handle summary as a JSON file.')
+  .option('--k6-handle-summary-json <path>','Output the K6 handle summary as a JSON file.')
   .action(run)
   .parse(process.argv);
 
@@ -46,8 +47,26 @@ async function run(...args) {
     console.error('Provide path to Postman collection');
     return;
   }
-  const options = args.pop();
+
+  let options = args.pop();
   const input = args.shift();
+
+  let cliOptions = {};
+  if (options.cliOptionsFile) {
+    try {
+      const cliOptionsFilePath = path.resolve(options.cliOptionsFile);
+      cliOptions = JSON.parse(await fs.readFile(cliOptionsFilePath, 'utf8'));
+    } catch (err) {
+      console.error(
+        '\x1b[31m',
+        `postman-to-k6 CLI options error - no such file or directory "${options.cliOptionsFile}"`
+      );
+      process.exit(1);
+    }
+  }
+
+  // Merge CLI configuration file with CLI parameters
+  options = Object.assign({}, cliOptions, options);
 
   // Convert
   let main, requests;
@@ -106,7 +125,7 @@ function translateOptions(options) {
       post: options.skipPost,
     },
     k6HandleSummary: {
-      json: options.k6HandleSummaryJson
+      json: options.k6HandleSummaryJson,
     },
   };
 }