+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/apideck-libraries/postman-to-k6/compare/v1.7.0...HEAD)
 
+### Added
+
+- Added option to use a file to pass all CLI options.
+
 ## [1.7.0] - 2020-09-01
 
 ### Added